Reorganise SearchResults and its Async variant
